Crate micromegas_analytics

Micromegas analytics: makes the telemetry data lake accessible and useful.

Modules

arrow_properties

Convert arrow array into Property instances

arrow_utils

Misc arrow utilities

async_block_processing

Processing of async events from thread event blocks

async_events_table

In-memory async events in arrow format

call_tree

Transforms thread events into call trees

delete

Removal of old data

dfext

Datafusion extensions

lakehouse

Module dedicated to the maintenance and query of materialized views

log_entries_table

In-memory log entries in arrow format

log_entry

Parsing of log entries from telemetry payload

measure

Parsing of metrics from telemetry payload

metadata

Access to the metadata stored in the relational database

metrics_table

In-memory metrics in arrow format

payload

Access to the raw binary telemetry payload

properties

UDF to convert properties to dictionary-encoded arrays Property-related User Defined Functions (UDFs) for DataFusion

record_batch_transformer

replication

bulk metadata & payload ingestion using Arrow

response_writer

Streams response for long requests

scope

Location in instrumented source code

span_table

In-memory call tree in arrow format

sql_arrow_bridge

Convert sqlx rows into arrow format

thread_block_processor

Parses thread event streams

time

Conversion between ticks and more convenient date/time representations