micromegas_analytics

Module delete

Removal of old data

Functions

delete_empty_processes

Deletes all empty processes from the data lake.

delete_empty_processes_batch

Deletes a batch of empty processes from the data lake. Returns true if there are more processes to delete, false otherwise.

delete_empty_streams

Deletes all empty streams from the data lake.

delete_empty_streams_batch

Deletes a batch of empty streams from the data lake. Returns true if there are more streams to delete, false otherwise.

delete_expired_blocks

Deletes all expired blocks from the data lake.

delete_expired_blocks_batch

Deletes a batch of expired blocks from the data lake. Returns true if there are more blocks to delete, false otherwise.

delete_old_data

Deletes all data older than a specified number of days from the data lake.