micromegas_analytics::lakehouse::partition_source_data

Trait PartitionBlocksSource

pub trait PartitionBlocksSource:
    Sync
    + Send
    + Debug {
    // Required methods
    fn is_empty(&self) -> bool;
    fn get_nb_blocks(&self) -> i64;
    fn get_max_payload_size(&self) -> i64;
    fn get_source_data_hash(&self) -> Vec<u8> ;
    fn get_blocks_stream<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = BoxStream<'static, Result<Arc<PartitionSourceBlock>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

A trait for providing blocks of source data for partitions.

Required Methods

fn is_empty(&self) -> bool

Returns true if there are no blocks in the source.

fn get_nb_blocks(&self) -> i64

Returns the number of blocks in the source.

fn get_max_payload_size(&self) -> i64

Returns the maximum payload size of the blocks in the source.

fn get_source_data_hash(&self) -> Vec<u8>

Returns a hash of the source data.

fn get_blocks_stream<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = BoxStream<'static, Result<Arc<PartitionSourceBlock>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns a stream of the source blocks.